LPS-185842 Extract ScopeFinder and ScopeMapper methods from CommerceServicePortalInstanceLifecycleListener to CommerceServiceScopeFinder and CommerceServiceScopeMapper
